changing emi_duration type to tinyint

+++ app/database/migrations/2014_01_12_123001_create_emi_plans_table.php
@@ -30,7 +30,7 @@ class CreateEmiPlansTable extends Migration {
 
             $table->integer(Emi\Entity::RATE);
 
-            $table->integer(Emi\Entity::DURATION);
+            $table->tinyInteger(Emi\Entity::DURATION);
             
             $table->string(Emi\Entity::METHODS)
             	  ->nullable();
